requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
boto3>=1.35.0
httpx>=0.27.0
python-dotenv>=1.0.0
langchain>=0.2.0
langchain-google-genai>=1.0.0
